Fixed errors and enabled ESLint back

+++ frontend/components/shared/StakeCard.tsx
@@ -80,7 +80,7 @@ const StakeCard = () => {
     useEffect(() => {
         if (isConnected && !isLoading && allowance?.result) {
             setNeedApproval(
-                allowance?.result < parseWBTC(wBTCAmount)
+                (allowance?.result as bigint) < parseWBTC(wBTCAmount)
             );
         }
     }, [allowance, wBTCAmount]);
@@ -88,7 +88,7 @@ const StakeCard = () => {
     useEffect(() => {
         if (isConnected && !isLoading && balance?.result) {
             setIsFundsSufficient(
-                balance?.result >= parseWBTC(wBTCAmount)
+                (balance?.result as bigint) >= parseWBTC(wBTCAmount)
             );
         }
     }, [balance, wBTCAmount]);
@@ -114,13 +114,13 @@ const StakeCard = () => {
 
     const expectedStBTCAmount = (_wBTCAmount : string) => {
         if (!exchangeRate?.result) { return "--" }
-        const result = parseWBTC(_wBTCAmount) * BigInt(10 ** EXCHANGE_RATE_DECIMALS) / exchangeRate?.result;
+        const result = parseWBTC(_wBTCAmount) * BigInt(10 ** EXCHANGE_RATE_DECIMALS) / (exchangeRate?.result as bigint);
         return formatStBTC(result * BigInt(10 ** 10)); // 10 = 18 - 8
     }
 
     const setMaxWBTCAmount = () => {
         if (!balance?.result) { return }
-        setWBTCAmount(formatWBTC(balance?.result))
+        setWBTCAmount(formatWBTC(balance?.result as bigint))
     }
 
     const stakeWBTC = () => {
@@ -210,7 +210,7 @@ const StakeCard = () => {
                 <div className="grid w-full items-center gap-4">
                     <div className="flex flex-col space-y-2">
                         <div className="flex flex-row mt-auto mb-auto">
-                            <Label >Your balance : { (isLoading || !balance?.result) ? "--" : formatWBTC(balance?.result)} wBTC</Label>
+                            <Label >Your balance : { (isLoading || !balance?.result) ? "--" : formatWBTC(balance?.result as bigint)} wBTC</Label>
                         </div>
                         <div className="flex flex-row mt-auto mb-auto">
                             <Input value={wBTCAmount} onChange={(e) => setWBTCAmount(e.target.value)} placeholder="Enter the amount of wBTC to stake" />
